Break ShredWAL into separate file and minor tweaks
